+++ src/main/kotlin/next/step/blackjack/domain/game/GameBoard.kt
@@ -5,7 +5,6 @@ import next.step.blackjack.domain.betting.BettingPlayer
 import next.step.blackjack.domain.betting.BettingPlayers
 import next.step.blackjack.domain.card.Card
 import next.step.blackjack.domain.card.Cards
-import next.step.blackjack.domain.card.GameCards
 import next.step.blackjack.domain.dealer.Dealer
 import next.step.blackjack.domain.player.Player
 import next.step.blackjack.domain.player.PlayerNames
