+++ src/main/kotlin/next/step/blackjack/domain/game/GameCards.kt
@@ -1,4 +1,8 @@
-package next.step.blackjack.domain.card
+package next.step.blackjack.domain.game
+
+import next.step.blackjack.domain.card.Card
+import next.step.blackjack.domain.card.CardFace
+import next.step.blackjack.domain.card.CardSymbol
 
 data class GameCards(private val cards: MutableList<Card>) {
