+++ src/main/kotlin/next/step/blackjack/domain/playercards/state/BurstState.kt
@@ -1,15 +1,15 @@
-package next.step.blackjack.domain.card.state
+package next.step.blackjack.domain.playercards.state
 
 import next.step.blackjack.domain.card.Cards
 import next.step.blackjack.domain.game.GameResult
 
-object BurstState : CardsState() {
+object BurstState : PlayerCardsState() {
 
-    override fun next(cards: Cards): CardsState {
+    override fun next(cards: Cards): PlayerCardsState {
         throw IllegalArgumentException("카드를 더 받을 수 없습니다.")
     }
 
-    override fun fight(other: CardsState): GameResult =
+    override fun fight(other: PlayerCardsState): GameResult =
         when (other) {
             BurstState -> GameResult.WIN
             else -> GameResult.LOSE
