+++ src/main/kotlin/next/step/blackjack/domain/playercards/state/UnfinishedState.kt
@@ -1,18 +1,18 @@
-package next.step.blackjack.domain.card.state
+package next.step.blackjack.domain.playercards.state
 
 import next.step.blackjack.domain.card.Cards
 import next.step.blackjack.domain.game.GameResult
 
-object UnfinishedState : CardsState() {
+object UnfinishedState : PlayerCardsState() {
 
-    override fun next(cards: Cards): CardsState = when {
+    override fun next(cards: Cards): PlayerCardsState = when {
         cards.isBlackjack() -> BlackjackState
         cards.isFinished() -> FinishedState
         cards.isBurst() -> BurstState
         else -> UnfinishedState
     }
 
-    override fun fight(other: CardsState): GameResult =
+    override fun fight(other: PlayerCardsState): GameResult =
         when (other) {
             BurstState -> GameResult.WIN
             BlackjackState -> GameResult.LOSE
