+++ src/test/kotlin/next/step/blackjack/domain/player/PlayersTest.kt
@@ -9,8 +9,8 @@ import next.step.blackjack.domain.card.Card
 import next.step.blackjack.domain.card.CardFace
 import next.step.blackjack.domain.card.CardSymbol
 import next.step.blackjack.domain.card.Cards
-import next.step.blackjack.domain.card.GameCards
 import next.step.blackjack.domain.dealer.Dealer
+import next.step.blackjack.domain.game.GameCards
 import next.step.blackjack.domain.game.GameResult
 import next.step.blackjack.domain.game.GameResults
 
@@ -197,9 +197,9 @@ class PlayersTest : DescribeSpec({
                 )
 
                 Players.of(setOf(player)).bet { BettingAmount.of(1000) } shouldBe
-                    BettingPlayers.of(
-                        setOf(BettingPlayer.of(player, BettingAmount.of(1000)))
-                    )
+                        BettingPlayers.of(
+                            setOf(BettingPlayer.of(player, BettingAmount.of(1000)))
+                        )
             }
         }
     }
